[BEAM-2877] Minor Go fixes for typos, proto changes and beamctl structure
