+++ sdks/go/cmd/beamctl/cmd/artifact.go
@@ -13,7 +13,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package main
+package cmd
 
 import (
 	"path/filepath"
